Integrated collapsiblePane in gui script + minor changes
Integrated cpane.py script into detect_chainsaw_gui.py
Added minimum window size (eq to initial window size)
